Create tests for login url and home

+++ database/migrations/2020_07_31_092808_create_role_user_table.php
@@ -13,10 +13,11 @@ class CreateRoleUserTable extends Migration
      */
     public function up()
     {
+        // php artisan migrate:rollback
         Schema::create('role_user', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('role_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('role_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->timestamps();
         });
     }
